+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - Add a widget for the instructor to upload thumbnail
+- Add video stats from configurable grafana backend
 
 ### Fixed